move-free-items-from-scheduler-py: Move 6 module-level free items out of scheduler.py
Pure mechanical relocation of 6 module-level free items out of
``scheduler.py``. Each item is byte-identical-relocated; no rename, no
signature change.

| Item | Target |
|---|---|
| ``EmbeddingBatchResult`` | ``managers/utils.py`` (alongside ``GenerationBatchResult``) |
| ``validate_dflash_request`` | ``speculative/dflash_utils.py`` |
| ``create_scheduler_watchdog`` | ``scheduler_components/invariant_checker.py`` |
| ``IdleSleeper`` | ``scheduler_components/idle_sleeper.py`` (new file) |
| ``is_health_check_generate_req`` | ``managers/utils.py`` (fixes reverse-import from ``tokenizer_manager``) |
| ``SenderWrapper`` | ``scheduler_components/output_sender.py`` (new file) |

Caller-site impact: only import-path rewrites. ``scheduler.py`` shrinks
by ~200 lines net.

A follow-up commit (``cleanup-scheduler-py-free-items``) handles the
non-mech cleanups (delete dead ``is_work_request``; rename
``SenderWrapper`` → ``SchedulerOutputSender`` to resolve the namesake
collision with ``multi_tokenizer_mixin.SenderWrapper``).

Refactor chain ID: move-free-items-from-scheduler-py
